pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "palaver"
version = "0.0.1"
description = "A Multi-Agent AI chatroom."
authors = [{ name = "Ruben Dedoncker", email = "ruben@dedoncker.dev" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fasta2a>=0.6.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langdiff>=0.2.0",
    "loguru>=0.7.3",
    "pydantic-ai-slim[a2a,anthropic,bedrock,cohere,google,groq,huggingface,mistral,openai,xai]>=1.64.0",
    "pydantic-settings>=2.13.1",
    "tomli-w>=1.2.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]

[project.scripts]
palaver = "palaver.app.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "twine>=6.2.0",
]

[tool.hatch.build.targets.wheel]
include = ["palaver/app/*", "palaver/ui/dist/*", "palaver/__init__.py"]

[tool.hatch.build.targets.sdist]
include = ["palaver/app/*", "palaver/ui/dist/*", "palaver/__init__.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["palaver"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "UP"]
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"palaver/app/main.py" = ["E402"]
